Disable cron and tweak parameters of i18n-download-tranlsations job.
In order to run this job from crowdin-go, I need it to not run itself
automatically, and also to accept the right parameters. While the ARCHIVEID
parameter is not currently used by the job, crowdin-go needs it to track which
archive was downloaded. (There is technically a possibility for a race
condition but I'm choosing to ignore this for now. Later on we can replace the
URL fetch from crowdin-go with a direct download of the archive from GCS.

Auditors: neha aasmund

Test plan: Crossed fingers.