Always include an "LTT Update" pipeline stage even if it's a no-op.
Jenkins UI doesn't like when the number of stages in the build pipeline changes
from build to build - in only shows the history until the last time the stages
changed. So having a stage that sometimes appears and sometimes doesn't causes
it to frequently reset.

Leaving the stage in and having it take zero time should hopefully make the UI
happier.

Auditors: neha

Test plan: Crossed fingers.