Add more reporting to emergency rollback job
Summary: Send slack messages at incremental stages of the emergency rollback job so users know the status of the job

Test Plan:
Have the `_alert` method only talk to slack if it's a DRY_RUN for now and manually check that the alerts are working properly when these changes have landed.
Update this method later when confirmed that alerts are working as intended!

Reviewers: jacqueline

Reviewed By: jacqueline

Subscribers: #sre

Differential Revision: https://phabricator.khanacademy.org/D55154