Call withSecrets() outside of webapp
Summary:
Following the landed changes on D55154, ran a test run and found that it failed: https://khanacademy.slack.com/archives/C096UP7D0/p1560644440051900

It seems that withSecrets() needs to be called within the workspace root, so this commit addresses that issue.

Test Plan:
Manually check the next dry run to see if the alerts are working properly when this change has landed. Then update this method later to run outside of dry runs when confirmed that alerts are working as intended.

Reviewers: jacqueline, csilvers

Reviewed By: jacqueline, csilvers

Differential Revision: https://phabricator.khanacademy.org/D55390