Add new `rollback-prod-postgres-migration` Groovy job for rolling back Postgres migrations
Summary:
Jira ticket: https://khanacademy.atlassian.net/browse/DIST-435

**Context**
As we continue to iterate on our Postgres schema, we'll want to make it easy to
roll back problematic migrations. This Groovy job (and the Jenkins job that
I'll create for it) should make it easy and safe to do so.

**Plan**
There are two parts to this ticket:
* (D56400) Adding instructions for the testing/staging/aborting process
* (This diff) Creating a Jenkins job to make it easy to rollback a specific migration

Test Plan: * Fingers crossed!

Reviewers: boris, csilvers, dhruv, kphilip

Differential Revision: https://phabricator.khanacademy.org/D56406