New job: Commit pofiles to webapp. (From i18n-download-translations)
Summary:
The bulk of the work in the i18n-download-translations job is being moved off
Jenkins and onto Kubernetes, however there's a fair bit of state and logic left
here just around cloning webapp and then committing pofiles to the
intl/translations subrepo. I'm going to leave this logic on Jenkins for the
moment, and just fetch the pofiles from GCS as a way to intermediate the two
parts of the job. Crowdin-go will invoke this job when the rest of the
download-translations work has been completed.

Test Plan:
- Crossed fingers - will test on live Jenkins.

Reviewers: neha

Subscribers: