Add support for running only a subset of tests, replacing `all_non_manual`.
Summary:
It used to be that when running tests, we could either run "all" the
tests (of a given size or less), or all tests except for those marked
@only_run_when_explicitly_specified_on_runtests_commandline.

The latter is only used by dev-e2e-test.groovy, and D57971 changes to
solving that some problem in Python.  So we can get rid of
all_non_manual.

Instead, I replaced it with a new mode, where you give it a git range,
and it figures out what files have changed in that git range and only
runs tests that depend on those files.  I use the new
deploy/should_run_tests.py tool to tell that.  Right now, nothing uses
that new mode, so everything runs in "all" mode like it did before
this change, but we'll be able to test the new mode once this is in.

Issue: none

Test Plan:
groovy jobs/webapp-test.groovy
groovy jobs/make-allcheck.groovy
groovy jobs/dev-e2e-test.groovy

But mostly fingers crossed.  You know how jenkins groovy development
is.

Reviewers: dhruv, jacqueline, benkraft

Reviewed By: dhruv

Subscribers: #infrastructure

Differential Revision: https://phabricator.khanacademy.org/D57975