Don't include --bad arg in a deploy-webapp rollback
Summary:
We don't create the git tag for a deploy until it ends with finishWithSuccess, so including the bad tag name when running rollback.py in deploy-webapp means that our rollback job will not recognize the tag and will fail. The rollback job works with just the good tag provided. The --bad arg is mostly there for marking a version bad in git (if it had already been merged to master).

Auditors: csilvers

Test Plan:
fingers crossed