Pass the new `--bad-dict` arg to rollback on abort.
Summary:
In 0501eae9a6607f we got rid of the `--bad` arg when calling
rollback.py -- which we do when there's an abort after set-default but
before finish -- because it didn't work properly with the new v3
version of deploy git-tags.  We really need the old style git-tags
that encode the entire version dict.

To resolve that, in D58220 I added a new flag to rollback.py that lets
you pass in the version-dict explicitly, in lieu of passing in a git
tag.  This is the commit that changes Jenkins to use that new flag.
As a result, we can specify the bad commit explicitly again, giving
more reliable rollbacks.

Test Plan: groovy jobs/deploy-webapp.groovy

Reviewers: jacqueline, benkraft

Reviewed By: jacqueline

Differential Revision: https://phabricator.khanacademy.org/D58221