Default BASE_REVISION to origin/master.
Summary:
This means that we'll turn on the "only run necessary tests" code for
phabricator diffs, leech-queue requests, and everything else that runs
webapp-test!  I've manually set BASE_REVISION to "" for allcheck, so
we do run all the tests at least once a day, and for the dev e2e
tests, which are flaky enough we want to run all of them just to be
safe.

Test Plan: Fingers crossed -- we'll see how it works when deployed!

Reviewers: benkraft, jacqueline

Reviewed By: jacqueline

Subscribers: Kai

Differential Revision: https://phabricator.khanacademy.org/D58365