Say what machine a given set of tests are running on.
This will help in debugging when tests fail and we want to try to
reproduce it exactly.  Putting the curl command inside `sh` means that
we can look at the logs for just this one log step and not have to go
through the gigs of testing output.

Auditors: jacqueline, benkraft

Test plan:
groovy jobs/webapp-test.groovy