Look for `go` binaries in `/usr/local/bin`, not `/snap`
Summary:

We install our go binaries in `/usr/local/bin` through a combination of
apt-get and symlinking (see `worker-setup.sh` in the aws-config repo for
details). But there's also go binaries available on jenkins in the
`/snap` directory, which auto updates through some other newfangled
ubuntu package management mechanism. We've pinned the version in
`/usr/local/bin` to 1.12, but the version in `/snap` auto upgraded to
1.13 last night, causing problems during deploy

This patch removes `/snap/bin` from the path that we define when running
a job on a worker and replaces it with `/usr/local/bin`. From the commit
history (0723d991), it looks like we added that entry just for go
binaries, so this shouldn't impact other jobs. Looking now on a jenkins
worker, there are google cloud binaries also defined in `/snap/bin`, but
I'm assuming we want to use the ones that are set in
`/home/ubuntu/google-cloud-sdk/bin`, which are set earlier in our PATH
anyhow

Test plan:
Try to deploy again, see no errors related to go version mismatch

Auditors: benkraft, Kai, csilvers