Ensure that 'prod-only' smoke tests actually run against prod
Summary:
While working on a new version of our graphql smoke test, I was confused
why the current version seemed to be failing on master, since I hadn't
seen anyone complain about it during deploys recently. Slack search
similarly came up empty and it was not on the list of QA skipped smoke
tests.

The issue turned out to be that we previously didn't pass the `--url`
flag in this groovy file when determine the list of smoke tests to run
on jenkins.  This meant that we'd always filter out the graphql smoke
test because it has a `prod_only` decorator [1] (even when we are
running tests against prod).

I think the test is failing on master because of outdate mobile native
queries uploaded for testing, but I'll double check that and make the
test pass before landing this diff.

[1]
https://github.com/Khan/webapp/blob/dhruv/web/response/end_to_end/graphql_smoketest.py#L37-39

Test Plan:
Cross fingers, push to jenkins, run the e2e test job and examine the
output to see if the two smoke tests with this decorator run as
expected (previously they didn't appear in the console log at all):
      - discussions/end_to_end/discussions_smoketest.py
      - web/response/end_to_end/graphql_smoketest.py

Reviewers: csilvers, benkraft

Subscribers: