Upload the graphql gateway config file when building webapp for deploys
Summary:
This patch adds a step to `build_webapp.groovy` to upload the latest
graphql gateway config file to GCS (by calling a soon to be added make
target in webapp).

The gateway schema config may change if any service changes the graphql
schema, so we want to it to trigger at least for those, but it's fast
enough (~2 seconds) for us to always call this for any deploy.

Test Plan: Say a prayer, deploy to jenkins and make sure the next few builds pass.

Reviewers: csilvers, benkraft, Kai

Subscribers: miguel

Differential Revision: https://phabricator.khanacademy.org/D60390