Truncate usernames for ZNDs
Now that we're being strict about the 63 characters thing, a lot of
people are hitting it.  And for good reason: you now only get 9
characters for your username and your version-name.  If your username is
10 characters, you're out of luck!

Now you're not: we truncate usernames to 8.  Ideally, we should
implement a better solution, but this will unblock people the quickest.

Test plan: ask Matt Morgan to run the job

Auditors: csilvers, #devops