Switch over to prod-VERSION URLs for smoke tests
I was hoping to keep the smoke tests in sync with everybody else, for
ease of testing.  But Brian's changes mean we need fastly in the loop
for them, so let's just use the new domain names today.  The only issue
is they don't do Google login, but neither do smoke tests, so it's fine.

Test plan: manually rerun a deploy smoke test job on the new URL.

Reviewers: csilvers, briangenisio