Upload a gateway config file when deploying a znd
Summary:
Whenever we deploy a new version of a service that changes the overall
composed schema, we need to upload a `gateway_config.json` file so that
the graphql-gateway can be aware of that version's schema. We do this
during deploys as part of the `build-webapp` job, but haven't done it
in the `deploy-znd` job so far.

This patch adds that step to `deploy-znd`. As with `build-webapp` we
only need to do for if the service in question changes the schema in
some way, but since it's fast we just do it for every znd deploy.

Jira ticket: None

Test Plan:
Cross fingers, land this on jenkins and deploy a znd. After deploying,
I'll check GCS[1] to make sure a file with the znd version is present.

[1]
https://console.cloud.google.com/storage/browser/ka-webapp/graphql-gateway/data_graph_configs/

Reviewers: csilvers, benkraft, #devops

Subscribers: davemcalpine