Set state to FAILURE when merging with master fails
Summary:
I think this fixes the most common cases. By setting the state of the jenkins job to FAILURE, buildmaster should persistently 'nag' users when their branch fails to merge master in correctly.

This does not address the direct usage of the deploy/merge-branches.groovy job (https://jenkins.khanacademy.org/job/deploy/job/merge-branches/). That job is only used by buildmaster directly, so any adjustment to it will have to happen there. It seems like the predominant (possibly only) usage of that job is when a deploy branch is picking up a leech, which doesn't seem to be what this task is really about.

Issue: INFRA-4811

Test Plan: Cross Fingers

Reviewers: benkraft, csilvers, #devops, avidal

Reviewed By: #devops, avidal

Differential Revision: https://phabricator.khanacademy.org/D64074