Initial tracing implementation
Summary:
In `vars/tracing.groovy` there's an implementation of a "sysdig-inspired" tracer, plus usage of it in webapp-test.

The gist is that you can decorate a job to have it emit span lines to standard
out, and then use an accompanying tool to consume the log and generate traces
from the spans.

Test Plan:
- merge
- cross fingers

Reviewers: davidbraley, csilvers

Reviewed By: davidbraley, csilvers

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D64293