Fix a few bugs with the tracing implementation
Summary:
The use of a first-argument `Map` to collect named params plus an optional
second argument `Span ctx` resulted in some odd method signature mismatching
when run on Jenkins. Notably, a `withSpan` call with a parent `Span` but no
`kwargs` would fail to find an implementation.

To simplify the method signature, I moved the `Map` to an optional third
argument that explicitly has to be a `Map` instead of a magical named param
slurping mechanism. Additionally, I removed the `null` default value from `ctx`
which is unneeded. All callers will specify a parent (or null) explicitly.

Test Plan:
- Run https://jenkins.khanacademy.org/job/misc/job/alexvidal-tracing-job/52/console
- Watch it pass

Reviewers: davidbraley, csilvers

Differential Revision: https://phabricator.khanacademy.org/D64444