Speed Up All Jenkins Jobs By About A Minute With This One Weird Trick!
For better or worse, in our jenkins jobs we typically force dependencies
to rebuild from scratch, rather than re-using whatever's on-disk.  This
is good because who knows what mess of a state the Jenkins workers are
in, but it takes a while.  We'd like to speed it up.

While looking into how to do that, I noticed it was running various
things multiple times.  I believe this is because `make -B deps` may end
up rebuilding things multiple times, because our fancy checksumming
stuff confuses Make so it can't deduplicate?  Whereas `fix_deps` just
invalidates all the checksums, such that everything gets reinstalled
exactly once, in theory.  It's not totally clear to me, but fix_deps is
significantly faster (1:42 rather than 2:56) on my machine, and they
should do the same thing, so I'll take it.

Test plan: crossed fingers

Reviewers: csilvers, avidal, davidbraley