Don't spin up more workers than we need for smoke test retries
If you're doing a retry, you may be running just a few tests.  But the
way we had it set up previously, we would already have started 20
workers, which meant we would need to wait for them to at least boot and
finish setup.  Now, we cap the number to what we will actually need, or
at least a guess that is in practice always right.

Test plan: crossed fingers

Reviewers: csilvers, avidal, davidbraley