WEB-1883: Remove jenkins deploy/build setup for react-render-server
Summary:
We don't need this anymore! I'm not sure if this is enough though, I suspect that there may be other Jenkins config that we need to change? But I'm not sure where we would do that.

We also need to delete the service from production but I don't think I have permission to do that:
https://khanacademy.atlassian.net/browse/WEB-1882

Issue: https://khanacademy.atlassian.net/browse/WEB-1883

Test Plan: Not sure how to test this!

Reviewers: #frontend-infrastructure, #sre, #devops, boris, davidbraley

Reviewed By: #sre, #devops, boris, davidbraley

Subscribers: davidbraley, csilvers, boris

Differential Revision: https://phabricator.khanacademy.org/D65233