Only run on Sauce if requested
Summary:
Running failures on sauce actually slows down the run quite a bit.
It seems [1] the majority of folks don't feel that's worth it, so let's
not do that!  I left it as an option, in case people want to do so for
manual runs.  We could also use that to enable it if you `sun: retry`
but people don't seem to want that either, at this time.

[1] https://khanacademy.slack.com/archives/C02NMB1R5/p1597268362123900

Test Plan: crossed fingers

Reviewers: csilvers, ruslan

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D65323