Share test-timing DBs across workers
Historically, we've just had each worker keep its own test-db, which
means it does its splitting based only on the previous jobs which have
used this worker to determine splits and analyze results.  This is never
great, but for unit tests it's good enough because we run `make check`
in worker build, so they always have *something* in the db.  (Plus, we
run many more unit test jobs, so it's more likely you hit a worker that
has something in the db.)

But for smoke tests it's terrible: it means you usually don't have any
splitting information to work from.  And it's worse, with D65328:
lexicographically closer tests probably have more correlated runtimes,
so if we are working with the default assumption that all tests take the
same amount of time, we'll be in real trouble.

Now, instead, we try to share the db.  It's all on a best-effort basis,
with no locking, but hopefully it ends up building up decent data over
time.

Test plan: will use "replay" to run some real jobs

Reviewers: csilvers

Subscribers: #devops