Remove no longer present build step to update bigquery jar
Summary:

In this patch [1], it looks like we removed the 'build_and_upload_jar'
target from the `dataflow/datastore_bigquery_adapter` job in favor of a
task to update a dataflow template.

To ensure that model changes are picked up we previously added a step to
deploys to call ` ./gradlew build_and_upload_jar`, which is now failing
since that task is gone.

I'm not sure if the right solution here is to replace this call with one
to `update_template`, but since I think it's okay for these model
definitions to be slightly out of date it seems safer to just delete
this step for now to unblock deploys and to have the data folks decide
whether to add back a call to `update_template` here.

[1] https://phabricator.khanacademy.org/D65224

Issue: None

Test plan:
Fingers crossed!

Auditors: tom