No longer make js deps for smoketests.
We needed them just for one place in our a11y smoketests, but I've
changed those tests to vendor the (small number of) js files required,
so we don't need to make deps anymore!  This should speed things up a
lot.

Note that we changed from fix_deps, which did a full re-download of
deps, to `python_deps`, which just does an incremental re-download of
python deps.  The reason for the former is that incremental updates of
js deps on jenkins has been flaky.  But incremental updates of python
deps have been solid, so we can move back to that now.  A big win!

Depends on: D65850

Issue: https://khanacademy.atlassian.net/browse/INFRA-5279

Test plan:
Mostly fingers crossed, will have to try it in real life!