Allow use of the new, dedicated 2nd-smoke-test workers for 2nd smoketests.
Summary:
We have created a new worker type that's just for use by the 2nd smoke
test (which is only ever run by the currently active deploy).  The
idea is that by dedicating machines to that task, we can speed it up
(particularly by not having to wait for new workers to start up if
they're all being used, which can take up to 5 minutes!)

This change supports using the new worker type.  It should be a noop
when we deploy it, until we change buildmaster to pass the new
boolean.

Issue: https://khanacademy.atlassian.net/browse/INFRA-5279

Test Plan:
groovy jobs/e2e-test.groovy
but as always, mostly fingers crossed

Reviewers: benkraft

Reviewed By: benkraft

Subscribers: #devops

Differential Revision: https://phabricator.khanacademy.org/D65888