Add a "none" mode to allow priming an e2e test machine.
Summary:
This mode doesn't run any tests at all -- it just does the git sync
and build-deps.  This mode is intended to be used with the second
smoke tests, which use a dedicated set of workers.  In that case,
running the e2e-test job early -- before we're ready to actually do
smoke-testing -- in "none" mode acts as a "priming" step: it makes
sure the worker nodes are active and it gets them up to date.  Then,
when the time comes to actually run the tests, it will re-use these
same nodes and they'll be ready to go.

(This works because the 2nd smoketests only run on the current deploy,
which runs in serial, so there's only ever one 2nd smoketest running
at a time.)

The actual machinery to do this "priming" build will happen in a later
commit.

Issue: https://khanacademy.atlassian.net/browse/INFRA-5279

Test Plan:
groovy jobs/e2e-test.groovy
but mostly fingers crossed.  I'll try running an e2e test manually
after this is landed!

Reviewers: benkraft

Reviewed By: benkraft

Subscribers: #devops

Differential Revision: https://phabricator.khanacademy.org/D65891