Move the 2nd-smoke-test priming to its own jenkins job.
Summary:
In testing, I realized that I needed to conditionally set
"allowConcurrentBuilds" based on whether we were priming or not.  That
was the tipping point for me to move the priming into its own job,
that is customized for what we actually need.

Issue: https://khanacademy.atlassian.net/browse/INFRA-5279

Test Plan:
groovy jobs/2ndsmoketest-priming.groovy
But mostly fingers crossed!  Will test when it's deployed.

Reviewers: benkraft

Reviewed By: benkraft

Subscribers: #devops

Differential Revision: https://phabricator.khanacademy.org/D65901