Make talking-to-buildmaster its own stage.
Talking to buildmaster can take a long time, since it waits 10 seconds
between retries on failure.  To aid in analysis of this, let's make
the buildmaster communication its own stage.  This will cause it show
up as its own row on the build-graph output I'm building.

e2e-test:16586 is an example of a build that had a lot of
buildmaster-waiting at the beginning and end.

Test plan:
Fingers crossed