Add REVISION_DESCRIPTION to merge-branches.
I am going to use this to as a key to associate different jenkins jobs
that are part of the same deploy.  (It's not perfect for that
purpose, which I may address later, but good enough for now.)

Right now all deploy jobs support REVISION_DESCRIPTION except
merge-branches.  This fixes that oversight!

Issue: none

Test plan:
groovy jobs/merge-branches.groovy