Add REVISION_DESCRIPTION to merge-branches.
Summary:
I am going to use this to as a key to associate different jenkins jobs
that are part of the same deploy.  (It's not perfect for that
purpose, which I may address later, but good enough for now.)

Right now all deploy jobs support REVISION_DESCRIPTION except
merge-branches.  This fixes that oversight!

Issue: none

Test Plan: groovy jobs/merge-branches.groovy

Reviewers: benkraft

Reviewed By: benkraft

Subscribers: #devops

Differential Revision: https://phabricator.khanacademy.org/D66172