Only build npm deps when we actually need them.
Summary:
build-webapp.groovy builds a lot of stuff, but -- perhaps surprisingly
-- all the js/npm building happens inside deploy_to_gcs.py.  So if
we're not deploying to gcs, we don' tneed to build any of our npm
files!

I did an audit of all the non-trivial scripts this job calls:
  deploy_to_gae.py
  deploy_to_gcs.py
  deploy/current_version.py
  deploy/chat_messaging.py

deploy_to_gae.py does a bunch of stuff, but none ruqires javascript.
The trickiest to check is that it calls deploy/build_files.py, but it
does that only with build_dynamic==True, and I confirmed that the kake
rules that this runs do not run any js.

deploy_to_gcs.py calls all sorts of node/js things, of course.

Test Plan: groovy jobs/build-webapp.groovy

Reviewers: #devops, benkraft

Differential Revision: https://phabricator.khanacademy.org/D66403