Run the ownership test before pushing new ownership data
We have some smoke-checks on the actual data file in here.  If they
fail, it's possible they need to be updated, or it's possible something
actually went wrong, but either way we should fix that before we try to
merge it into every single deploy!

Test plan: fingers crossed

Auditors: dhruv, csilvers