Send a survey when a deployment gets aborted
Summary:
Infrastrucutre would like to better understand why deployments get aborted. Because the decision is almost always made by a human, we have no automated way to gather information on this subject. To that end, a simple survey can help us get actual data to better understand these decisions.
We'll start with a simple google survey, sent asynchronously to the branch deployer and andyone who was cc'd on the deployment. We recognise that the person who aborted a deployment isn't necessarily in this set all of the time, but it is frequently the case that they are. We'll also send ourselves a copy so we can follow up if needed.
We further recognise that a google survey is less likely to gather results than something like an in-slack survey, but requires significantly less work to get up and running and start getting results from. It's reasonable to assume that our engineers want the deployment system to be as quick and problem free as possible, so may well be willing to go a little more above-and-beyond to get us results in this case. If that turns out to be an incorrect assumption, we can swing back and replace this with a slack questionaire or similar.

Issue: INFRA-5338

Test Plan: Fingers Crossed! (also some hacky tests outlined in the Jira)

Reviewers: #devops, csilvers, benkraft

Reviewed By: #devops, benkraft

Subscribers: avidal

Differential Revision: https://phabricator.khanacademy.org/D66872