Don't pass in the `--url` flag in server mode.
runsmoketests.py uses `--url` to decide whether to include the
prod-only smoketests or not, but runtests_server.py uses the `--prod`
flag instead, and complains it doesn't know about the `--url` flag.
No problem, boblem.

Auditors: benkraft

Test plan:
Ran this manually and got the server to run successfully, and still
saw it running:
   testing/runsmoketests.py --timing-db=genfiles/test-info.db --url=https://www.khanacademy.org -n --just-split -j80
to do the splits.