Allow trying the client/server architecture for webapp-test
Summary:
This ports the logic that is already in e2e-test.groovy, and makes it
available in webapp-test.groovy as well.  I don't enable the new
architecture by default yet, but I'm interested in playing around with
it to see how it performs.

One nice characteristic of the old, split-based method of doing tests
is we could try to keep all the js tests (say) on a small number of
machines, and still be confident in doing equal splits.  In the
server-based scheme we can't really do that, so all our workers will
have to pay the overhead of starting up jest, possibly several times
(since we don't distribute all the js tests at once).  So this is one
thing in particular to be on the lookout for when exploring this new
system.

Issue: none, but kinda https://khanacademy.atlassian.net/browse/INFRA-5458

Test Plan:
I successfully ran this groovy script on
   https://jenkins.khanacademy.org/job/deploy/job/webapp-test/71297

Reviewers: #devops, benkraft

Reviewed By: #devops, benkraft

Differential Revision: https://phabricator.khanacademy.org/D66975