Use the test-client/server architecture always, for e2e tests.
I ran in client/server mode for the last week or so, with no
complaints.  An analysis of running time shows they are comparable.
(But in real life I believe client/server is better; an analysis of
some of the slow runs shows I'm not yet properly handling the case the
last test-worker comes online after all tests are running, which slows
down the run a lot in that case.)

To generate the comparison data, I ran
   mkdir /tmp/html
   gsutil -m rsync gs://ka-jenkins-perf/ /tmp/html/
   grep -l e2e-test /tmp/html/*.html | sh -c 'while read f; do grep -B5 "Running tests" "$f" | grep timeRangeRelativeToBuildStart | sed "s/[^(]*(/`date -r $f +%s\ `/" | cut -d")" -f1 ; done' > /tmp/html/plot
and then
   % gnuplot
   > set arrow nohead from 1603411418,0 to 1603411418,2000 ls 5
   > plot '/tmp/html/plot'

Issue: https://khanacademy.atlassian.net/browse/INFRA-5514

Test plan:
None