Use `failfast` to cancel workers rather than shared state.
Our previous way of telling workers there's no work to do -- perhaps
because the test-server couldn't start up successfully -- was to set a
global var.  This works, but is not ideal because if a test-server is
still being provisioned (the gce machine is still starting up), it
won't see the global var until after provisioning is done, which means
pointless waiting around.

So now I try a different technique: using `failfast` to cancel the
workers in that situation.  I don't know for sure that `failfast` will
cancel a provision-in-progress, but I'm hoping so!

It's a little bit of a hack because we want to "fail" fast even on
success, if all tests were run before all workers were provisioned.  I
use a custom exception to hopefully make it clear what's going on.

As a side effect of this, we no longer support FAILFAST for its
original purpose of running until the first test fails.  But that's
ok; nobody ever used that option on jenkins anyway.

Issue: none

Test plan:
I ran this on jenkins with success, using "replay" of
   https://jenkins.khanacademy.org/job/deploy/job/e2e-test/18031/
I then did another reply where I changed
   sh("env HOST=${serverIP} testing/runtests_server.py --smoketests ${exec.shellEscapeList(runSmokeTestsArgs)}")
to
   sh("false")
and the job failed very quickly, without provisioning all the
test-workers.