Remove some unused parameters from webapp-test.groovy
Summary:
I'm attempting to clean up the public API of some of our Jenkins jobs
to make it easier to refactor them.  In particular I'm getting rid of
job params that seemed like they'd be useful at the time, but in
practice are never used.

I found the candidate list like so:
   % gcloud --project khan-internal-services compute ssh jenkins@jenkins-server
   $ cd jobs/deploy/jobs/webapp-test/builds
   $ perl -nle 'm,<name>(.*)</name>, and $name = $1; m,<value>(.*)</value>, and print "$name: $1"' */build.xml  | sort | uniq -c | sort -nr | less

This showed the following params that never had a non-default value
set in the 500 most recent webapp-test jobs, to wit:
    506 TEST_FILE_GLOB: *_test.py
    506 NUM_RETRIES: 0
    506 JOB_DISPLAY_NAME: webapp-test
    506 FAILFAST: false
    506 CLEAN: some
    504 NUM_WORKER_MACHINES: 20

This diff gets rid of the first 4.  I kept CLEAN because even though
it hasn't been used recently, I know it's been used historically.  And
I kept NUM_WORKER_MACHINES because that's something I actually want to
play around with.  (And the nightly allcheck job is going to use a
different value for this anyway.)

Test Plan: groovy jobs/webapp-test.groovy

Reviewers: #devops, benkraft

Reviewed By: #devops, benkraft

Differential Revision: https://phabricator.khanacademy.org/D67211