No longer sync khan-linter as part of webapp-test.
Summary:
Back in the day, we had linter tests that called into
devtools/khan-linter directly.  But now that webapp owns its own
linter, it doesn't need to do that anymore!  So we can get rid of the
special-case code here for that.

I double-checked this by doing
   git grep khan-linter
   git grep devtools
in webapp-root, and not finding any relevant hits.

Test Plan: Fingers crossed

Reviewers: #devops, benkraft

Reviewed By: #devops, benkraft

Differential Revision: https://phabricator.khanacademy.org/D67266