Update webapp-test to support multiple test-runners on one machine.
Summary:
I adapted the existing, battle-tested code in e2e-test.groovy for this
purpose.  For now, the default number of test-runners per worker is
still 1, but we'll be upping it (to 3) to match the optimal number
according to my testing.

I could have done this a lot easier: by passing in `-j3` to
runtests.py to tell it to use 3 CPUs.  I didn't do this for a few
reasons:

1) We've found it really helpful for debugging, that jenkins runs
tests in a deterministic order (that it emits).  This lets us
reproduce bugs caused by leaky tests.

2) In the future, we'll have many test-clients running: not just
runtests.py but a jest for js, gradle for kotlin, etc.  Some have the
equivalent of `-j3` and some don't.  The approach in this commit means
it doesn't matter if they do or not.

3) Even if all our test-clients supported the equivalent of `-j3`,
it's not so efficient to run:
    runtests.py -j3; jest -j3; gradlew -j3; go test -cpu 3; etc.
This means that if one CPU finishes python tests (say) before the
others do, it sits idle until the other CPUs finish, instead of going
straight on to start running the jest tests.

And of course, it's nice for the e2e-test code and webapp-test code to
match more closely.  Maybe one day I'll try to unify them.

While in the area, I did some renaming to make e2e-test and
webapp-test use the same terminology.  I also took the opportunity to
get rid of old reference to "test splitting," which we don't really do
anymore, or "determining tests," which is now subsumed into "test
serving."

Issue: none

Test Plan:
I ran this code using the jenkins "replay" feature, at
   https://jenkins.khanacademy.org/job/deploy/job/webapp-test/72860

Reviewers: #devops, benkraft

Reviewed By: #devops, benkraft

Differential Revision: https://phabricator.khanacademy.org/D67283