Move junit xml generation to the workers.
This is the beginning of the plumbing to change webapp-test to use the
junit xml files as the canonical source of test-result data, rather
than our test-db.  We'll eventually use the xml for both pass/fail
status and timing.  But this change is much less ambitious: it just
changes the place we convert from test-db to xml from being on the
main test-server process, to being on each of the test-worker
processes.

This will allow us to start using test-clients that generate junit xml
but don't know anything about our internal, bespoke test-db format.
As long as the client can give junit xml back to the test-server (via
stash/unstash, in our case), then everything will work.

My big worry is that multiple workers will produce conflicting
filenames.  I use subdirs to try to avoid this.

Test plan:
I used jenkin's "replay" functionality to run this at
   https://jenkins.khanacademy.org/job/deploy/job/webapp-test/72894/
I made it so this would have test-failures, to verify that the junit
plugin is properly reading our new-format junit xml.