Accept dataflow-batch in build scripts.
Invokes `deploy/deploy_to_dataflow.py` if `dataflow-batch` is passed in.
As this is a gradlew job, we run this sequentially with the existing Kotlin
jobs.

This change should be a no-op as our `should_deploy.py` changes have not yet
been pushed in.

Issue: https://khanacademy.atlassian.net/browse/INFRA-5606

Test plan:
<see below>