Use our own stashing rather than stash(), for test-results.
Summary:
We've been seeing periodic failures in the stash() command.  It's
unclear why, the java exceptions are not illuminating.  It may be that
it's being interrupted by an exception, but I don't see any exceptions
happening at that time in the flow table.  It may also be there's just
a transient communication error trying to connect to the stash-server.

In any case, let's avoid the problem altogether by using our own stash
mechanism, which I happen to have incorporated into test_server.py.
While that mechanism was intended to stash .xml files, it actually
works fine to stash any type of file -- it doesn't do the junit
analysis it would do if you uploaded .xml files, but that's fine in
our case, we do that anlaysis later in the jenkins job anyway.

Test Plan:
I re-ran a failed test-run of mine, at
   https://jenkins.khanacademy.org/job/deploy/job/webapp-test/74350
It properly reported (to slack) the 1 failing test, and did not
complain about tests not being run or anything like that.

Reviewers: #devops, benkraft

Reviewed By: #devops, benkraft

Subscribers: marksandstrom

Differential Revision: https://phabricator.khanacademy.org/D67571