Use `catchError` to fail the build on failing tests.
People have noticed that sometimes the slack message will indicate
tests failed, while jenkins will be all green.  That is because there
are some types of test problems that only the slack-message-summarizer
notices.

I've changed slack-message-summarizer (in webapp) to exit with a
non-zero rc if it notices problems.  But we don't want to fail the
jenkins job right away when that happens: it has more work to do after
that!  Instead we use `catchError` to mark the build as failed but
continue execution.

Auditors: benkraft

Test plan:
Fingers crossed, will run it on jenkins