Make sure we unstash test-info.db into a clean location.
I'm seeing this on jenkins-server:
   lrwxrwxrwx 1 jenkins jenkins       18 Nov 30 14:46 /var/lib/jenkins/jobs/deploy/jobs/webapp-test/workspace/test-info.db -> ../../test-info.db

I don't know how this can happen.  My only theory is that on one of
the workers, somehow genfiles/test-info.db is a symlink.  And
`unstash` doesn't delete the symlink but follows it instead, and then
`stash` stashes the symlink (and unstashes it as a symlink on
jenkins-server).  This theory doesn't really make sense -- I confirmed
that the test-worker image doesn't have a symlink in
genfiles/test-info.db -- but I don't have a better one.

In any case, sticking in these rm's first won't hurt anything.  And
maybe they'll help, who knows.

Auditors: benkraft

Test plan:
Fingers crossed