Add a "separate" test script for phabricator diffs.
We run our test suite in two different situations: when doing a
deploy, and whenever a phabricator diff is created or updated.  They
work the same way, and have historicalily been run by the same jenkins
job.

This commit is the first step to change that, allow for 2 jenkins jobs
with, potentially, 2 different workflows.  For now they are identical
-- I'm committing the new groovy file as a symlink -- but we could
change that if we wanted to.  For now the main difference is the two
are separated out in the jenkins ui, which has its plusses and minuses
but I think overall is a win.

Auditors: benkraft

Test plan:
None