Switch over to using the native test-runners!
Summary:
This is a major rewrite of webapp-test.groovy.  We still start the
runtests-server as before, but we no longer use the python
test-wrappers for running go tests, js tests, and lints.  Instead, we
use the native test-clients for them.

A lot of the difficulty here is not actually moving to the different
test-runners, but handling the test-results; as part of this change we
use junit xml files as the canonical source of test output, not a
python-specific pickle file.

Another great source of difficulty is properly handling errors, with
the goal of being able to see the red dot on the pipeline-view page
when test-runs fails, but still have parallel tests continue.  I
believe I came up with a system that is self-evident in how it works,
but I'm sure jenkins will find a way it won't work in some situations!

This new technique seems to take about the same amount of time on
shorter test-runs, but is significantly faster on longer ones.  For a
full test-run on `csilvers`, the old method takes about 42 minutes
while the new one takes about 33.

Change phabricator-test.groovy instead, leave webapp-test alone for now.

This will let us test this new framework in pieces:
Step 1) Use the new framework for test-runs due to `arc diff` (this change)
Step 2) Use the new framework for test-runs during deploys (future change)

Test Plan:
I ran many tests on this diff over time.  The last was on a
branch that had a bug that caused it not to generate junit files in
some cases:
   https://jenkins.khanacademy.org/job/deploy/job/webapp-test/75141/

(none new)

Reviewers: #devops, benkraft

Reviewed By: #devops, benkraft

Differential Revision: https://phabricator.khanacademy.org/D67673