Use the new test-job everywhere!  Not just for phabricator diffs.
In our 24 hours of testing we didn't see any problems, so we're trying
the new job everywhere!

In fact we've already gotten rid of use of the phabricator-test
jenkins job, so I'm getting rid of the phabricator-test.groovy file as
part of this diff too.

Auditors: benkraft

Test plan:
Fingers very crossed!